feat: DIsplay Refine Stat on Thymeleaf View

+++ src/main/java/com/thestat/thestat/domain/PlayerStat.java
@@ -45,7 +45,7 @@ public class PlayerStat
 
     private int rebound;
 
-    private double gamesPlayed;
+    private int gamesPlayed;
 
     private int totalPointMade;
 }
